requirements.txt
mcp-server>=0.1.0
Pillow>=9.0.0
